pyproject.toml
[project]
name = "tp-lumina-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.128.0",
    "firebase-admin>=7.1.0",
    "google-cloud-firestore>=2.23.0",
    "google-cloud-storage>=3.8.0",
    "jinja2>=3.1.6",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.20",
    "uvicorn>=0.39.0",
]
